# cmd/server: concurrently search instead of serial iterations