This change will spread CPU usage across available cores and increase memory usage where possible, which offers better search performance.

These searches are still not nearly fast enough for me, but they are steady progress we can build from going forward.

```
// before (v0.17.1)
// Go 1.15.7
√ cmd/server % go test -bench BenchmarkSearch -run BenchmarkSearch -v
goos: darwin
goarch: amd64
pkg: github.com/moov-io/watchman/cmd/server
BenchmarkSearch__Addresses
BenchmarkSearch__Addresses-16      	      22	  50782847 ns/op
BenchmarkSearch__BISEntities
BenchmarkSearch__BISEntities-16    	     216	   5415917 ns/op
BenchmarkSearch__DPs
BenchmarkSearch__DPs-16            	    1855	    611191 ns/op
BenchmarkSearch__SDNs
BenchmarkSearch__SDNs-16           	      94	  12140552 ns/op
BenchmarkSearch__SSIs
BenchmarkSearch__SSIs-16           	     602	   1941815 ns/op
PASS
ok  	github.com/moov-io/watchman/cmd/server	15.228s
```
```
// after (sync.WaitGroup)
// Go 1.15.7
√ cmd/server % go test -bench BenchmarkSearch -run BenchmarkSearch -v
goos: darwin
goarch: amd64
pkg: github.com/moov-io/watchman/cmd/server
BenchmarkSearch__Addresses
BenchmarkSearch__Addresses-16      	      33	  34522864 ns/op
BenchmarkSearch__BISEntities
BenchmarkSearch__BISEntities-16    	    1161	   1021614 ns/op
BenchmarkSearch__DPs
BenchmarkSearch__DPs-16            	    3991	    296631 ns/op
BenchmarkSearch__SDNs
BenchmarkSearch__SDNs-16           	     394	   3042463 ns/op
BenchmarkSearch__SSIs
BenchmarkSearch__SSIs-16           	    2161	    505961 ns/op
PASS
ok  	github.com/moov-io/watchman/cmd/server	14.404s
```
```
// Go 1.16.3
√ cmd/server % go test -bench BenchmarkSearch -run BenchmarkSearch -v
go version
goos: darwin
goarch: amd64
pkg: github.com/moov-io/watchman/cmd/server
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkSearch__Addresses
BenchmarkSearch__Addresses-16      	      38	  27957405 ns/op
BenchmarkSearch__BISEntities
BenchmarkSearch__BISEntities-16    	    1180	   1014248 ns/op
BenchmarkSearch__DPs
BenchmarkSearch__DPs-16            	    4315	    277961 ns/op
BenchmarkSearch__SDNs
BenchmarkSearch__SDNs-16           	     392	   3023502 ns/op
BenchmarkSearch__SSIs
BenchmarkSearch__SSIs-16           	    2431	    503114 ns/op
PASS
ok  	github.com/moov-io/watchman/cmd/server	14.318s
```