@adamdecaf 2021-04-28
Thanks! The first code block in the PR are the "before" results. We're seeing some significant improvements with the most naive concurrent approach.